Rename cstruct file
