+++ dissect/database/ese/ntds/objects/c_dnsnode.py
